+++ tests/serverpod_test_module/serverpod_test_module_server/lib/serverpod_test_module_server.dart
@@ -3,3 +3,4 @@ export 'src/endpoints/ignored_endpoint.dart';
 export 'src/endpoints/unauthenticated.dart';
 export 'src/generated/endpoints.dart';
 export 'src/generated/protocol.dart';
+export 'src/test_module_module.dart';